Encapsulate save/load in viewmodel
